refactor: RestaurantTimeFactory 중간에 추상클래스를 추가
템플릿 메소드 패턴 적용하여 중복코드 제거
