+++ src/main/java/com/posomo/saltit/domain/restaurant/domain/restaurantTime/factory/RestaurantBreakTimeFactory.java
@@ -2,7 +2,6 @@ package com.posomo.saltit.domain.restaurant.domain.restaurantTime.factory;
 
 import com.posomo.saltit.domain.restaurant.dto.RestaurantTimeDto;
 import com.posomo.saltit.domain.restaurant.entity.Restaurant;
-import com.posomo.saltit.domain.restaurant.entity.core.enums.Day;
 import com.posomo.saltit.domain.restaurant.entity.manyToOne.RestaurantBreakTime;
 import com.posomo.saltit.domain.restaurant.entity.manyToOne.RestaurantTime;
 import com.posomo.saltit.domain.restaurant.entity.manyToOne.RestaurantTimeDay;
@@ -10,16 +9,12 @@ import java.util.List;
 import java.util.UUID;
 
 
-public class RestaurantBreakTimeFactory implements RestaurantTimeFactory {
+public class RestaurantBreakTimeFactory extends AbstractRestaurantTimeFactory{
 
     @Override
-    public RestaurantTime createTime(Restaurant restaurant, RestaurantTimeDto timeDto) {
-        UUID timeId = UUID.randomUUID();
-        RestaurantTime timeFk = new RestaurantTime(timeId);
-        List<RestaurantTimeDay> days = Day.toRestaurantTimeDayList(timeFk,
-                timeDto.getDays());
+    protected RestaurantTime build(UUID id, Restaurant restaurant, List<RestaurantTimeDay> days, RestaurantTimeDto timeDto) {
         return RestaurantBreakTime.builder()
-                .id(timeId)
+                .id(id)
                 .timeTo(timeDto.getTimeTo())
                 .timeFrom(timeDto.getTimeFrom())
                 .restaurant(restaurant)
